Внесены изменения по уроку "Разделение ответственности" спринта 5
