+++ MovieQuiz/Models/QuizQuestion.swift
@@ -1,0 +1,14 @@
+//
+//  QuizQuestion.swift
+//  MovieQuiz
+//
+//  Created by Aleksey Yakushev on 10.03.2023.
+//
+
+import Foundation
+
+struct QuizQuestion {
+    let image: String
+    let text: String
+    let correctAnswer: Bool
+}
