+++ MovieQuiz/Models/QuizResultsViewModel.swift
@@ -1,0 +1,16 @@
+//
+//  QuizResultsViewModel.swift
+//  MovieQuiz
+//
+//  Created by Aleksey Yakushev on 10.03.2023.
+//
+
+import Foundation
+
+// для состояния "результат квиза"
+
+struct QuizResultsViewModel {
+    let title: String
+    let text: String
+    let buttonText: String
+}
