+++ MovieQuiz/Models/QuizStepViewModel.swift
@@ -1,0 +1,17 @@
+//
+//  QuizStepViewModel.swift
+//  MovieQuiz
+//
+//  Created by Aleksey Yakushev on 10.03.2023.
+//
+
+import Foundation
+import UIKit
+
+// для состояния "вопрос задан"
+
+struct QuizStepViewModel {
+    let image: UIImage
+    let question: String
+    let questionNumber: String
+}
